+++ Project.swift
@@ -60,6 +60,8 @@ let project = Project(
             sources: [
                 "MessagesExtension/Tests/**",
                 "MessagesExtension/Sources/Presentation/**",
+                "MessagesExtension/Sources/Board/GameBoardRenderModel.swift",
+                "MessagesExtension/Sources/Board/GameBoardRenderModelBuilder.swift",
             ],
             dependencies: [
                 .package(product: "ULS_CoreGame"),